pyproject.toml
[build-system]
requires = [
    "setuptools",
]
build-backend = "setuptools.build_meta"

[project]
name = "django-calendar"
version = "0.1"
description = "Calendar with Django Models"
dependencies = [
    "Django>=5.1.1",
    "python-dateutil>=2.9.0.post0",
]
readme = "README.md"
requires-python = ">= 3.10"
authors = [
    {name = "Eduardo Stefanello", email = "eduardo@stefanello.cc"},
]
classifiers = [
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
]

[project.urls]
Homepage = "https://github.com/dudustefanello/django_calendar/"
